Add Toggle component and remove unnecessary code
